Implement support for the X86_64 Overflow Flag (OF)
Introduce support for emulating the Overflow Flag on the ppc64le
backend. This is done in a similar fashion to the Carry Flag (CF)
by lazily evaluating the flag when it is checked (i.e. by a conditional
jump).

Instead of the jump table approach used by the CF code, this uses
a single comparison to branch to the shared flag calculation code
(regardless of width) for the last operation type (add, sub, etc.)
and then shifts the result according to the operation width using
the shift constants in ppc64le::LastFlagOpData.