arch/ppc64le/codegen: Major refactor of assembler
Majorly refactor the ppc64le assembler and some of the code that
uses it.

Overview
========

The old assembler interface was simple and functional, but it greatly
limited our ability to inspect and modify the generated instructions
after they were generated but before they were emitted, since the
assembler would immediately insert them into a code buffer.
Thus, passes that wished to modify the generated instructions either
needed to store out-of-band data about them that they could use to
re-emit them to the code buffer after modification, or disassemble
the buffer with capstone. The former approach was taken by the
relocation code, but it was much clunkier than it needs to be.

The new assembler interface offers significantly more flexibility
by not immediately emitting the instructions into a code buffer,
but instead storing the instruction's information in a newly
introduced instruction stream. The instruction stream contains
the instruction type, all operands to the instruction, and a
callback that can be used to finally emit the instruction to a code
buffer.

Case Study: Relocation
======================

Using the new interface, code can walk the instruction stream
after instruction generation and inspect/modify the instruction
arbitrarily before it actually gets emitted. The relocation code
has been updated to take advantage of this - it no longer requires
storing the instruction type and operands out-of-band and then
overwriting the instructions in the code buffer during its pass.
Instead, the relocation information is simply attached to entries
of the instruction stream and the relocation pass walks the through
the stream until a relocation is encountered. Once encountered,
the relevant parts of the instruction are modified in-place
according to the relocation information.

With this new approach, instructions are only emitted into a
code buffer once - at the end of codegen_ppc64le::translate() -
after all intermediate passes have modified the stream where
necessary.

Implementation
==============

A major design goal of the refactor was to preserve the simple
API exposed by the assembler class while still adding powerful
inspection and modification facilities for intermediate passes.
Specifically, the pattern of calling one assembler method to
generate one instruction (e.g. assembler::addi(rt, ra, si)) is
left unchanged.

To add the new functionality without modifying the API, each
assembler instruction method was modified to invoke its underlying
code emitting function (e.g. xl_form) through the EMIT_INSN macro,
rather than invoking it directly. EMIT_INSN accepts, among others,
a lambda callback which contains the calls to the underlying emitting
functions, and adds it to the assembler's instruction stream.

Some template and macro tricks are (ab)used to capture the lambda's
state (i.e. the instruction's operands provided by the caller) in
an array of std::variant types which are stored along with the lambda
in the instruction stream entry. Compared to capturing these
values directly in the lambda, this approach allows code to inspect
the values and modify them before calling the lambda - the primary
design goal of the refactor.

Code that wishes to access the operands of an instruction stream
entry can do so by using the `insn_get<n, Op>(insn)` helper template
which further abuses some template metaprogramming techniques to
provide a type-correct reference to the operand at index `n` of
the instruction stream entry `insn`.

To attempt to isolate the assembler logic from the newly introduced
template and macro hacks, a new header, util/magic.h was created
to store their implementations. Verbose comments were added to
attempt to explain the functionality of each hack.

Result
======

The result of the refactor is a signficantly more powerful API for
the assembler that has simplified the relocation code and paves the way
for future optimization passes.

As a result of the large amount of template and macro metaprogramming,
the compile time for the ppc64le codegen code has increased quite a bit,
but ultimately this is a small price to pay for the flexibility gained.

In spite of the increase in compile times though, benchmarks show that
runtime performance of the ppc64le backend has actually increased by a
non-negligible amount. I haven't investigated the reason for this in
depth, but I suspect it has to do with the fewer code region allocations
and resizes that we perform - since now we only allocate a code region
once after the instruction stream is finalized, so we know the exact
size to allocate and don't need to perform any grow or shrink operations
on it.