Rewrite simple_placement_allocator to be even simpler
The old simple_placement_allocator was a piece of garbage and
contained multiple bugs (lack of MAP_FIXED flag, incorrect resizing,
etc.) and was in dire need of a rewrite. The new version is the
simplest possible implementation of a placement allocator. Responsibility
of mmaping the target region is moved to the code that constructs
the simple_placement_allocator (in this case execution_context), and
the entire managed region is mmap'd at once instead of lazily mapped
as-needed.