Implement support for PUSH/POP and stack allocation
Implement support for basic X86_64 stack instructions, modeled in
llir as loads/stores with update. In addition to updating the ppc6le
codegen backend to support loads/stores with update and memory operands
(in the case of push [mem]), execution_context was modified to allocate
a stack for the translated process to use.