arch/ppc64le/codegen: Add support for local label relocations
Soup up the ppc64le relocation system to add support for declaring and
resolving local labels like a real assembler. This removes the need for
hardcoding branch offsets in things like the lazy flag evaluation macros.

To reduce the boilerplate required to extract and inspect function
arguments in the relocation resolution code, a new set of annotated
types were added to codegen_types.h which, when used as parameters to an
assembler function, can convey more specific information about their
usage in the instruction. B and BC have been updated to use these new
annotated types, which allows the relocation code to modify their
relative offset fields without hardcoding a table of per-instruction
argument information.

The annotated types are implemented as empty enums with explicitly
specified underlying stdint.h types. Doing it this way instead of with
typedef/using allows the types to be recognized as distinct types by the
C++ type system instead of mere aliases. This means that they can be
included in std::variants along with their underlying type without
causing issues. This is kind of an abuse of enums, but it works quite
well.