Support translating jumps to non-local targets
This commit introduces a major overhaul to introduce support for jumping
to non-local targets, e.g. for function calls. Since retrec starts off by
only translating the entrypoint routine, a mechanism for lazily
translating other referenced routines has been introduced.

When a non-local call/jump is encountered by codegen_ppc64le, a call
to trap_patch_call or trap_patch_jump will be emitted to the instruction
stream, along with the target virtual address encoded as a raw constant
in the instruction stream, and a few NOPs (if necessary). At run-time,
when one of these traps is hit, runtime_context_ppc64le::execute() will
return to dynamic_recompiler's execution loop with status_code::UNTRANSLATED.

dynamic_recompiler will then call into backend-specific code to retrieve
the address of the untranslated access, then it will translate the
routine using translate_elf_function or the newly introduced
translate_raw_code_block. The latter is used as a fallback for when the
target virtual address doesn't lie within an ELF function in the target
binary, and it will only translate up until the first branch-type
instruction.

After translating the region, dynamic_recompiler will pass the host
address of the newly translated function to the codegen so that it can
patch in a call/jump instead of the trap. The codegen will also rewind
the instruction pointer so that when execution is resumed it will
execute the newly patched instructions.

Finally, dynamic_recompiler will re-enter the translated code and
execution will resume.