Implement support for the IMUL instruction
Implement support for the IMUL instruction (all three forms) in the
x86_64 llir lifter and the ppc64le codgen backend.

On the llir side, a new Alu::Op was added for IMUL, and a new FlagArr
field `flags_undefined` was added to signal that the states of some
flags may be clobbered without restoring them.

On the ppc64le codegen backend, support to the llir$alu code was added
for emitting all 3 forms of IMUL, and honoring the `flags_undefined`
field to avoid unnecessary flag save/preserve instructions from being
emitted.

Also, support for lazily calculating the overflow/carry flag (they're
the same for IMUL) was added for the new operation type. The previous
carry flag lazy evaluation was rewritten from scratch to more closely
resemble the overflow calculation code. The new IMUL flag calculation
code was added to a fixed helper that the overflow/carry flag macros
call when the last flag operation type is IMUL.