Build our own capstone instead of using system's
By building our own copy of capstone instead of using the system's, we
can enable DIET mode which results in an additional ~30% performance
increase (!) at the cost of removing features that we didn't even use
(except for in a couple of sanity checks which have been removed).

Add capstone as a git submodule along with a simple cmake wrapper for
building it with the required options. Also clean up the CMakeLists file
a bit to set the default CMAKE_BUILD_TYPE and fix building in Release
mode by ensuring that asserts don't have any necessary side effects,
since they don't get evaluated in RELEASE mode.