Implement DEC, OR, SETcc, HLT instructions
Implement support for DEC, OR, SETcc, and HLT in the x86_64 llir lifter
and the ppc64le codegen backend. The former 3 are all represented as
llir::Alu instructions and the latter is represented as the newly
introduced llir::Alu::Interrupt::ILLEGAL, denoting an instruction that
may not be executed from userspace code.

CPUID was also added to the x86_64 llir lifter, but no support has been
added to the ppc64le codegen backend yet.