Implement support for CMOVcc instructions
Add support for CMOVcc instructions to x86_64 llir_lifter. New
llir::Insn field `condition`, was added to allow these instructions to
be represented.

On the ppc64le backend, any instruction with a condition field is
guarded by a conditional branch that skips the translation if the
condition is not true. This has the advantage of working with any
existing llir instruction without special handling and is probably the
fastest way we can implement CMOVcc on ppc64le anyways.