arch/ppc64le/codegen: Fix loads/stores to <64-bit register operands
Previously macro$loadstore unconditionally assumed that the base/index
registers have a mask of Full64. As expected, this breaks horribly when
storing to smaller registers, like `mov [eax], X`. Update it to allocate
a temporary register and use macro$move_register_masked to copy the
value into the temporary before dereferencing it. Also add some juggling
to support PRE/POST Updates when this temporary register is used.

As an aside, macro$loadstore is getting pretty unwieldy at this point
with a bunch of divergent code paths and somewhat hard to follow
register management. It should probably be broken down/partially
rewritten soon.