Support REP/REPE/REPNE instruction prefixes
Introduce support for REP* prefixes in the x86_64 llir lifter, llir, and
the ppc64le codegen backend.

In llir, REP* is represented with the new llir::Qualification class.
llir::Qualification is used as a catch-all for specifying any special
behavior associated with an instruction. The previous
llir::Insn::conditional field, used for CMOV*, has also been migrated to
use the new Qualificaiton class.

In the ppc64le codegen backend, qualifications are handled in
translate() before the instruction is dispatched, and in the case of
llir::Qualification::Repeat, after the dispatch as well.